Переделано на общий layout для треков
