+++ src/app/tracks/layout.tsx
@@ -3,31 +3,19 @@
 import styles from "./page.module.css";
 import { Nav } from "@/components/nav/nav";
 import { Search } from "@/components/search/search";
-import { Player } from "@/components/player/player";
 import { Filter } from "@/components/filter/filter";
 import { ContentPage } from "@/components/content/contentpage";
 import { Sidebar } from "@/components/sidebar/sidebar";
-import { useEffect, useMemo } from "react";
-import { useAppDispatch, useAppSelector } from "@/store/store";
-import {
-  fetchFavoritesTracks,
-  setTracksList,
-} from "@/store/features/playListSlice";
+import { Player } from "@/components/player/player";
+import classNames from "classnames";
+import { useAppSelector } from "@/store/store";
 
-export default function Favorites() {
-  const classNames = require("classnames");
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
   const { currentTrack } = useAppSelector((state) => state.playList);
-  const { favoritesList } = useAppSelector((state) => state.playList);
-  const { error } = useAppSelector((state) => state.playList);
-  const dispatch = useAppDispatch();
-
-  useEffect(() => {
-    dispatch(setTracksList(favoritesList));
-    if (favoritesList.length === 0) {
-      dispatch(fetchFavoritesTracks());
-      dispatch(setTracksList(favoritesList));
-    }
-  }, [dispatch, favoritesList]);
 
   return (
     <div className={styles.wrapper}>
@@ -37,9 +25,9 @@ export default function Favorites() {
           <div
             className={classNames(styles.mainCenterblock, styles.centerblock)}>
             <Search />
-            <h2 className={styles.centerblockH2}>Мои треки</h2>
+            {children}
             <Filter />
-            <ContentPage error={error} />
+            <ContentPage />
           </div>
           <Sidebar />
         </main>
