+++ src/components/search/search.module.css
@@ -1,5 +1,5 @@
 .centerblockSearch {
-  width: 100%;
+  width: 1109px;
   border-bottom: 1px solid #4e4e4e;
   margin-bottom: 51px;
   display: -webkit-box;
