+++ src/components/sidebar/sidebar.module.css
@@ -1,5 +1,5 @@
 .mainSidebar {
-  max-width: 418px;
+  width: 418px;
   padding: 20px 90px 20px 78px;
 }
