+++ tsconfig.json
@@ -22,6 +22,6 @@
       "@/*": ["./src/*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "middleware.js", "src/Track.test.tsx", "src/ProgressBar.test.tsx", "src/NavBar.test.tsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "middleware.js", "src/Track.test.tsx", "src/ProgressBar.test.tsx", "src/NavBar.test.tsx", "src/app/tracks/page.tsx"],
   "exclude": ["node_modules"]
 }
